services/mishka-admin-backend/pyproject.toml
[project]
name = "mishka-admin-backend"
version = "0.1.0"
description = "Backend for Mishka AI Admin Panel"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyjwt>=2.8.0",
    "aiogram>=3.0.0",
    "python-dotenv>=1.0.0",
    "email-validator>=2.0.0",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.28.0",
    "aio-pika>=9.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "httpx"
]
